tests/wpt/web-platform-tests/css/css-multicol/parsing/column-count-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: parsing column-count with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-count">
<meta name="assert" content="column-count supports the full grammar 'auto | <integer>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("column-count", "auto");

test_valid_value("column-count", "1");
test_valid_value("column-count", "234");
</script>
</body>
</html>
